refactor test running into separate file
