+++ src/extension.ts
@@ -2,6 +2,7 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
 import * as ctest from "./test_discovery";
+import { run_tests } from "./test_runner";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -30,7 +31,7 @@ export function activate(context: vscode.ExtensionContext) {
     "Run",
     vscode.TestRunProfileKind.Run,
     (request, token) => {
-      ctest.run_tests(controller, log_channel, false, request, token);
+      run_tests(controller, log_channel, false, request, token);
     }
   );
