+++ src/test/package_test.sh
@@ -9,8 +9,11 @@ expected_files="[Content_Types].xml
 extension/CHANGELOG.md
 extension/images/logo.png
 extension/LICENSE.txt
+extension/out/extension_helpers.js
 extension/out/extension.js
+extension/out/test_details.js
 extension/out/test_discovery.js
+extension/out/test_runner.js
 extension/package.json
 extension/README.md
 extension.vsixmanifest"
